% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_flow.R
\docType{data}
\name{ex_flow}
\alias{ex_flow}
\title{Traded volume of land between countries}
\format{A data frame with 10866 rows and 6 variables:
\describe{
  \item{O}{name of exporting country}
  \item{D}{name of importing country}
  \item{vegetable}{volume of land associated with trading vegetables, in ha}
  \item{fruit}{volume of land associated with trading fruits, in ha}
  \item{wheat}{volume of land associated with trading wheats, in ha}
  \item{soybean}{volume of land associated with trading soybeans, in ha}
}}
\source{
\url{http://fao.org/faostat/}
}
\usage{
data(ex_flow)
}
\description{
A dataset containing trade data between countries and traded volumes for each 4 category.
}
\keyword{datasets}
