% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_node.R
\docType{data}
\name{ex_node}
\alias{ex_node}
\title{country attributes}
\format{A data frame with 154 rows and 8 variables:
\describe{
  \item{country}{name of exporting country}
  \item{x}{longitude of the center of a country}
  \item{y}{latitude of the center of a country}
  \item{pop_total}{total number of population}
  \item{gdpc}{Gross Domestic Product per capita, in dollar}
  \item{area_cultivation}{total volume land for cultivation use, in ha}
  \item{water_total}{total volume of usable water, in cubic meter}
  \item{income_level}{5 levels by income}
}}
\source{
\url{http://fao.org/faostat/}
}
\usage{
data(ex_node)
}
\description{
A dataset containing 154 countries who participate in the trade and related attributes.
}
\keyword{datasets}
